% Generated by roxygen2 (4.0.1): do not edit by hand
\name{b2g_to_gsc}
\alias{b2g_to_gsc}
\title{Blast2GO to GSC}
\usage{
b2g_to_gsc(file, organism = "my-organism")
}
\arguments{
\item{file}{Tab separated text: gene_id GO_accession GO_description}

\item{organism}{The name of your organism}
}
\value{
A GSEA GeneSetCollection object
}
\description{
converts a tab-delimited text file to a GOstat gsc object
}
\examples{
GO_gsc = b2g_to_gsc(file="myBlast2GO.annot",organism="myOrganism")
}
\keyword{Blast2GO}
\keyword{GO,}
\keyword{GSEA,}

