% Generated by roxygen2 (4.0.1): do not edit by hand
\name{significant_terms}
\alias{significant_terms}
\title{significant_terms}
\usage{
significant_terms(GO_OBJ, cutoff)
}
\description{
\code{significant_terms} extracts the significant GO terms from a GOstat
results object
}
\details{
Details of function go here
}

