% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_GO}
\alias{test_GO}
\title{test_GO}
\usage{
test_GO(genes, ontology, gsc, direction = "over", universe, pval = 0.01)
}
\description{
\code{test_GO} returns GO terms overrepresented in a gene set
}
\details{
Use GO annotations and GOstats in R to quickly test for over-represented GO terms
}

