% Generated by roxygen2 (4.0.1): do not edit by hand
\name{b2g_to_gsc}
\alias{b2g_to_gsc}
\title{b2g_to_gsc}
\usage{
b2g_to_gsc(file, organism = "my-organism")
}
\description{
\code{b2g_to_gsc} converts a tab-delimited text file to a GOstat gsc object
}
\details{
Input file format: geneId GOId geneDesc (tab separated)
}

